Implemented authentication + route guarding to the app
Right now, you can sign up and sign in in the app and you aren't allowed
to go to the Form page if you're not signed in. (I'll remove this in
the next commit - I only implemented it as a proof of concept.)

When signing up, it creates an account on Firebase, however, it doesn't
make a database entry. I need to fix it so it does that.

I also need to make it look better, eventually.

Contributes to #19