Fixed book details breaking on page reload
Before, it was trying to get the id for the book onInit, but instead of
doing that, I'm just saving the id of the book in LocalStorage (which
doesn't get wiped on refresh) when we click on "Details".

Doing it this way lets us avoid querying the book list every time as
well.