Added new environment specifically for testing and added screenshots in tests
I made a completely different Firebase project with the same data in
order to not use our limited writes/reads with our main project. It calls
it with npm run e2e or npm test, and it fetches from the other
Firebase database.

I also implemented the functionality to take screenshots for e2e tests

Contributes to #37, #36, #21, #41